+++ Bencodex/Types/Integer.cs
@@ -4,6 +4,7 @@ using System.Diagnostics.Contracts;
 using System.Globalization;
 using System.IO;
 using System.Numerics;
+using System.Security.Cryptography;
 using System.Text;
 
 namespace Bencodex.Types
@@ -63,6 +64,21 @@ namespace Bencodex.Types
         /// <inheritdoc cref="IValue.Type"/>
         [Pure]
         public ValueType Type => ValueType.Integer;
+
+        /// <inheritdoc cref="IValue.Fingerprint"/>
+        [Pure]
+        public Fingerprint Fingerprint
+        {
+            get
+            {
+                // If the byte representation is compact enough, use it as a digest too.
+                // If it's longer than 20 bytes, make a SHA-1 hash for digest.
+                byte[] bytes = Value.ToByteArray();
+                IReadOnlyList<byte> digest =
+                    bytes.Length <= 20 ? bytes : SHA1.Create().ComputeHash(bytes);
+                return new Fingerprint(Type, EncodingLength, digest);
+            }
+        }
 
         /// <inheritdoc cref="IValue.EncodingLength"/>
         [Pure]